support for gapcursor in figure
